+++ adapters/AppLovin/AppLovinAdapter/OpenBidding/GADMRTBAdapterAppLovinInterstitialRenderer.m
@@ -73,7 +73,7 @@
   GADMAdapterAppLovinExtras *extras = self.adConfiguration.extras;
   self.sdk.settings.muted = extras.muteAudio;
 
-  [self.interstitialAd showOver:[UIApplication sharedApplication].keyWindow andRender:self.ad];
+  [self.interstitialAd showAd: self.ad];
 }
 
 - (void)dealloc {
@@ -99,7 +99,7 @@
 #pragma mark - Ad Load Delegate
 
 - (void)adService:(ALAdService *)adService didLoadAd:(ALAd *)ad {
-  [GADMAdapterAppLovinUtils log:@"Interstitial did load ad: %@", ad.adIdNumber];
+  [GADMAdapterAppLovinUtils log:@"Interstitial did load ad: %@", ad];
 
   GADMRTBAdapterAppLovinInterstitialRenderer *parentRenderer = self.parentRenderer;
   parentRenderer.ad = ad;