+++ adapters/Facebook/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Facebook iOS Mediation Adapter Changelog
 
+#### Next Version
+- Adapter now returns a non-zero `mediaContent` aspect ratio once the media view is rendered in a view.
+
+#### Version 5.5.1.0
+- Verified compatibility with FAN SDK 5.5.1.
+
 #### Version 5.5.0.0
 - Verified compatibility with FAN SDK 5.5.0.
 - Now requires Google Mobile Ads SDK version 7.46.0 or higher.