+++ adapters/I-Mobile/CHANGELOG.md
@@ -1,4 +1,5 @@
 # i-mobile Mediation Adapter for Google Mobile Ads SDK for iOS
 
-## Version x.y.z.a
-- Under development
+## Version 2.0.29.0
+- Initial release!
+- Added support for banner, interstitial, and native ads.